Feature/books

+++ service/category/pg.go
@@ -67,6 +67,12 @@ func (s *pgService) Delete(_ context.Context, p *domain.Category) error {
 		}
 		return err
 	}
+
+	// delete all associated books
+	errDeleteBooks := s.db.Where("category_id = ?", p.ID).Delete(domain.Book{}).Error
+	if errDeleteBooks != nil {
+		return errDeleteBooks
+	}
 	return s.db.Delete(old).Error
 }
